+++ vpn_manager/__init__.py
@@ -1,1 +1,3 @@
-from .__main__ import *
+from . import peers
+
+__all__ = ["peers"]
